requirements.txt
pypdf2==3.0.1
python-docx==1.1.0
nltk==3.8.1
google-generativeai==0.3.2
psycopg2-binary==2.9.9
pgvector==0.2.4
python-dotenv==1.0.0
tqdm==4.66.1
tenacity==8.2.3
